Error Handling

+++ sample/src/main/java/com/ezdk/sample/Controller/ProductServiceController.java
@@ -1,5 +1,7 @@
-package com.ezdk.sample.Product;
+package com.ezdk.sample.Controller;
 
+import com.ezdk.sample.Middleware.ProductNotfoundException;
+import com.ezdk.sample.Model.Product;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -25,6 +27,7 @@ public class ProductServiceController {
 
     @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable("id") String id) {
+        if(!productRepo.containsKey(id))throw new ProductNotfoundException();
         productRepo.remove(id);
         return new ResponseEntity<>("Product is deleted successfully", HttpStatus.OK);
     }
